TOB-6019 Refaktor pakker og bestillingservice

+++ src/test/kotlin/no/nav/sokos/skattekort/listener/KafkaListener.kt
@@ -34,7 +34,7 @@ object KafkaListener : BeforeSpecListener {
     fun getKakfaConfig() = KafkaConfig()
 
     fun createKafkaTopic(vararg topics: String) {
-        val newTopics = topics.map { topic -> NewTopic(topic, 1, 1.toShort()) }
+        val newTopics = topics.map { topic -> NewTopic(topic, 1, 1) }
         adminClient.createTopics(newTopics)
     }
